refactor(skill): split CLI command domains
