+++ framework/core/src/python/kungfu/cli/commands/_skill/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""Responsibility-owned implementations behind ``kungfu skill``."""
